Updated the function load_retriever() so it does not concatenate index_root twice in a row.
The function concatenates index_root and index_name into a single index path, and then pass it to the method RAGMultiModalModel.from_index() when the path already exists.

However, from_index() already does the concatenation internally as well, so index_root will be concatenated twice in a row. This tpyically results in a file-not-found error.